make `resolve_features` a stand alone function
