+++ src/cargo/core/resolver/dep_cache.rs
@@ -4,10 +4,12 @@ use std::rc::Rc;
 
 use log::debug;
 
-use crate::core::{Dependency, PackageId, PackageIdSpec, Registry};
+use crate::core::interning::InternedString;
+use crate::core::{Dependency, FeatureValue, PackageId, PackageIdSpec, Registry, Summary};
 use crate::util::errors::CargoResult;
 
-use crate::core::resolver::types::Candidate;
+use crate::core::resolver::types::{Candidate, ConflictReason};
+use crate::core::resolver::{ActivateResult, Method};
 
 pub struct RegistryQueryer<'a> {
     pub registry: &'a mut (dyn Registry + 'a),
@@ -172,5 +174,238 @@ impl<'a> RegistryQueryer<'a> {
         self.cache.insert(dep.clone(), out.clone());
 
         Ok(out)
+    }
+}
+
+/// Returns all dependencies and the features we want from them.
+pub fn resolve_features<'b>(
+    parent: Option<PackageId>,
+    s: &'b Summary,
+    method: &'b Method<'_>,
+) -> ActivateResult<(
+    HashSet<InternedString>,
+    Vec<(Dependency, Vec<InternedString>)>,
+)> {
+    let dev_deps = match *method {
+        Method::Everything => true,
+        Method::Required { dev_deps, .. } => dev_deps,
+    };
+
+    // First, filter by dev-dependencies.
+    let deps = s.dependencies();
+    let deps = deps.iter().filter(|d| d.is_transitive() || dev_deps);
+
+    let reqs = build_requirements(s, method)?;
+    let mut ret = Vec::new();
+    let mut used_features = HashSet::new();
+    let default_dep = (false, Vec::new());
+
+    // Next, collect all actually enabled dependencies and their features.
+    for dep in deps {
+        // Skip optional dependencies, but not those enabled through a
+        // feature
+        if dep.is_optional() && !reqs.deps.contains_key(&dep.name_in_toml()) {
+            continue;
+        }
+        // So we want this dependency. Move the features we want from
+        // `feature_deps` to `ret` and register ourselves as using this
+        // name.
+        let base = reqs.deps.get(&dep.name_in_toml()).unwrap_or(&default_dep);
+        used_features.insert(dep.name_in_toml());
+        let always_required = !dep.is_optional()
+            && !s
+                .dependencies()
+                .iter()
+                .any(|d| d.is_optional() && d.name_in_toml() == dep.name_in_toml());
+        if always_required && base.0 {
+            return Err(match parent {
+                None => failure::format_err!(
+                    "Package `{}` does not have feature `{}`. It has a required dependency \
+                     with that name, but only optional dependencies can be used as features.",
+                    s.package_id(),
+                    dep.name_in_toml()
+                )
+                .into(),
+                Some(p) => (
+                    p,
+                    ConflictReason::RequiredDependencyAsFeatures(dep.name_in_toml()),
+                )
+                    .into(),
+            });
+        }
+        let mut base = base.1.clone();
+        base.extend(dep.features().iter());
+        for feature in base.iter() {
+            if feature.contains('/') {
+                return Err(failure::format_err!(
+                    "feature names may not contain slashes: `{}`",
+                    feature
+                )
+                .into());
+            }
+        }
+        ret.push((dep.clone(), base));
+    }
+
+    // Any entries in `reqs.dep` which weren't used are bugs in that the
+    // package does not actually have those dependencies. We classified
+    // them as dependencies in the first place because there is no such
+    // feature, either.
+    let remaining = reqs
+        .deps
+        .keys()
+        .cloned()
+        .filter(|s| !used_features.contains(s))
+        .collect::<Vec<_>>();
+    if !remaining.is_empty() {
+        let features = remaining.join(", ");
+        return Err(match parent {
+            None => failure::format_err!(
+                "Package `{}` does not have these features: `{}`",
+                s.package_id(),
+                features
+            )
+            .into(),
+            Some(p) => (p, ConflictReason::MissingFeatures(features)).into(),
+        });
+    }
+
+    Ok((reqs.into_used(), ret))
+}
+
+/// Takes requested features for a single package from the input `Method` and
+/// recurses to find all requested features, dependencies and requested
+/// dependency features in a `Requirements` object, returning it to the resolver.
+fn build_requirements<'a, 'b: 'a>(
+    s: &'a Summary,
+    method: &'b Method<'_>,
+) -> CargoResult<Requirements<'a>> {
+    let mut reqs = Requirements::new(s);
+
+    match *method {
+        Method::Everything
+        | Method::Required {
+            all_features: true, ..
+        } => {
+            for key in s.features().keys() {
+                reqs.require_feature(*key)?;
+            }
+            for dep in s.dependencies().iter().filter(|d| d.is_optional()) {
+                reqs.require_dependency(dep.name_in_toml());
+            }
+        }
+        Method::Required {
+            all_features: false,
+            features: requested,
+            ..
+        } => {
+            for &f in requested.iter() {
+                reqs.require_value(&FeatureValue::new(f, s))?;
+            }
+        }
+    }
+    match *method {
+        Method::Everything
+        | Method::Required {
+            uses_default_features: true,
+            ..
+        } => {
+            if s.features().contains_key("default") {
+                reqs.require_feature(InternedString::new("default"))?;
+            }
+        }
+        Method::Required {
+            uses_default_features: false,
+            ..
+        } => {}
+    }
+    Ok(reqs)
+}
+
+struct Requirements<'a> {
+    summary: &'a Summary,
+    // The deps map is a mapping of package name to list of features enabled.
+    // Each package should be enabled, and each package should have the
+    // specified set of features enabled. The boolean indicates whether this
+    // package was specifically requested (rather than just requesting features
+    // *within* this package).
+    deps: HashMap<InternedString, (bool, Vec<InternedString>)>,
+    // The used features set is the set of features which this local package had
+    // enabled, which is later used when compiling to instruct the code what
+    // features were enabled.
+    used: HashSet<InternedString>,
+    visited: HashSet<InternedString>,
+}
+
+impl Requirements<'_> {
+    fn new(summary: &Summary) -> Requirements<'_> {
+        Requirements {
+            summary,
+            deps: HashMap::new(),
+            used: HashSet::new(),
+            visited: HashSet::new(),
+        }
+    }
+
+    fn into_used(self) -> HashSet<InternedString> {
+        self.used
+    }
+
+    fn require_crate_feature(&mut self, package: InternedString, feat: InternedString) {
+        self.used.insert(package);
+        self.deps
+            .entry(package)
+            .or_insert((false, Vec::new()))
+            .1
+            .push(feat);
+    }
+
+    fn seen(&mut self, feat: InternedString) -> bool {
+        if self.visited.insert(feat) {
+            self.used.insert(feat);
+            false
+        } else {
+            true
+        }
+    }
+
+    fn require_dependency(&mut self, pkg: InternedString) {
+        if self.seen(pkg) {
+            return;
+        }
+        self.deps.entry(pkg).or_insert((false, Vec::new())).0 = true;
+    }
+
+    fn require_feature(&mut self, feat: InternedString) -> CargoResult<()> {
+        if feat.is_empty() || self.seen(feat) {
+            return Ok(());
+        }
+        for fv in self
+            .summary
+            .features()
+            .get(feat.as_str())
+            .expect("must be a valid feature")
+        {
+            match *fv {
+                FeatureValue::Feature(ref dep_feat) if **dep_feat == *feat => failure::bail!(
+                    "cyclic feature dependency: feature `{}` depends on itself",
+                    feat
+                ),
+                _ => {}
+            }
+            self.require_value(fv)?;
+        }
+        Ok(())
+    }
+
+    fn require_value(&mut self, fv: &FeatureValue) -> CargoResult<()> {
+        match fv {
+            FeatureValue::Feature(feat) => self.require_feature(*feat)?,
+            FeatureValue::Crate(dep) => self.require_dependency(*dep),
+            FeatureValue::CrateFeature(dep, dep_feat) => {
+                self.require_crate_feature(*dep, *dep_feat)
+            }
+        };
+        Ok(())
     }
 }
